Finished breaking apart phase space and velocity models into separate modules. Full test suite now passes
