+++ halotools/empirical_models/phase_space_models/__init__.py
@@ -1,3 +1,5 @@
-from .phase_space_models import *
+from .nfw_phase_space import *
+from .trivial_phase_space import *
+
 from .profile_models import *
 from .velocity_models import *
